guts of a SortedMap, based on a balanced BST
TreeMap doesn't have a floor concept, so can't be used, and it needed
to be rewritten/replaced anyway. The implementation is complete enough
to attack day 5, though most of the methods are unimplemented, and just
exit with a message to STDERR.

The balancing is crude, and can be improved, but it works.